# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is API document for UAMS backend server",
        "title": "Backend API",
        "contact": {
            "name": "Mr. Khai"
        },
        "version": "1.0"
    },
    "host": "http://iot.hcmue.space:8079",
    "basePath": "/v1",
    "paths": {
...
        "/v1/package_accesses": {
            "get": {
                "description": "find all package access log",
                "produces": [
                    "application/json"
                ],
                "summary": "Find Package Access log",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/models.PackageAccess"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/package_accesses/area_id/{area_id}": {
            "get": {
                "description": "find all package access log by  and AreaID",
                "produces": [
                    "application/json"
                ],
                "summary": "Find Package Access log by AreaID",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/models.PackageAccess"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/package_accesses/area_id/{area_id}/period/{from}/{to}": {
            "get": {
                "description": "find all package access log by AreaID and TimeRange",
                "produces": [
                    "application/json"
                ],
                "summary": "Find Package Access log by AreaID and TimeRange",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/models.PackageAccess"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/package_accesses/package_id/{id}": {
            "get": {
                "description": "find all package access log by Package ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Find Package Access log by Package ID",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/models.PackageAccess"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/package_accesses/package_id/{id}/area_id/{area_id}": {
            "get": {
                "description": "find all package access log by Package ID and AreaID",
                "produces": [
                    "application/json"
                ],
                "summary": "Find Package Access log by Package ID and AreaID",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/models.PackageAccess"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/package_accesses/package_id/{id}/area_id/{area_id}/period/{from}/{to}": {
            "get": {
                "description": "find all package access log by Package ID and AreaID and TimeRange",
                "produces": [
                    "application/json"
                ],
                "summary": "Find Package Access log by Package ID and AreaID and TimeRange",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/models.PackageAccess"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/package_accesses/package_id/{id}/period/{from}/{to}": {
            "get": {
                "description": "find all package access log by Package ID and TimeRange",
                "produces": [
                    "application/json"
                ],
                "summary": "Find Package Access log by Package ID and TimeRange",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/models.PackageAccess"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/package_accesses/period/{from}/{to}": {
            "get": {
                "description": "find all package access log by TimeRange",
                "produces": [
                    "application/json"
                ],
                "summary": "Find Package Access log by TimeRange",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/models.PackageAccess"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.PackageAccess": {
            "type": "object",
            "properties": {
                "area_id": {
                    "type": "string"
                },
                "group": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "package_id": {
                    "type": "string"
                },
                "random": {
                    "type": "string"
                }
            }
        },
...
        "utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "errorMsg": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        }
    }
}